creates bubble chart in a separate state and
refactores http request, data manipulation and d3 chart drawing
into separate services
