+++ src/app/gameCharts/bubbleChart_directive.js
@@ -1,0 +1,13 @@
+export const bubbleChart = [
+  'BubbleChartDataServices',
+
+  (BubbleChartDataServices) => {
+  return {
+    link: function (scope, element, attrs) {
+      scope.view = false;
+      console.log('attr', attr)
+
+    }
+  }
+}
+];
